% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shownotes.R
\name{get_show_notes}
\alias{get_show_notes}
\title{Search for Setlist}
\usage{
get_show_notes(apikey, showdate = NULL)
}
\arguments{
\item{apikey}{apikey}

\item{showdate}{the show setlist in YYYY-MM-DD format}
}
\value{
the selected show's show notes
}
\description{
Search for Setlist
}
\examples{
\dontrun{
BigCypressNYE <- get_show_notes(apikey = "<apikey>", showdate = "1999-12-31")
}
}
