% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setlist.R
\name{get_setlist}
\alias{get_setlist}
\title{Search for Setlist}
\usage{
get_setlist(apikey = getOption("phishr_key"), showdate = NULL)
}
\arguments{
\item{apikey}{Your key for the Phish.net API. This can also be stored
as an option using \code{options('phishr_key') <- 'your_api_key'}.}

\item{showdate}{the show setlist in YYYY-MM-DD format}
}
\value{
the selected show from search
}
\description{
Search for Setlist
}
\examples{
\dontrun{
BigCypressNYE <- get_setlist(apikey = "<apikey>", showdate = "1999-12-31")
}

}
