% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishnet_get_setlist.R,
%   R/phishnet_get_setlist_2.R, R/phishnet_get_show_rating.R,
%   R/phishnet_get_shownotes.R
\name{pn_get_setlist}
\alias{pn_get_setlist}
\alias{pn_get_setlist2}
\alias{pn_get_show_rating}
\alias{pn_get_show_notes}
\title{Query the Phish.Net API}
\usage{
pn_get_setlist(apikey = getOption("phishnet_key"), showdate = NULL)

pn_get_setlist2(apikey = getOption("phishnet_key"), showdate = NULL)

pn_get_show_rating(apikey = getOption("phishnet_key"), showdate = NULL)

pn_get_show_notes(apikey = getOption("phishnet_key"), showdate = NULL)
}
\arguments{
\item{apikey}{Your key for the Phish.net API. This can also be stored
as an option using \code{options('phishnet_key') <- 'your_api_key'}. See details
for more information on how to obtain one.}

\item{showdate}{the show setlist in YYYY-MM-DD format}
}
\value{
the selected show's rating, setlist, or notes as a \code{data.frame}.
}
\description{
The \code{phish.net} API contains a vast amount of information.
The functions here are specific to each type of query that can be sent to the
API - all are prefixed with \code{pn_*}
}
\details{
Accessing the Phish.Net API requires an API key. These can be obtained
by following the steps at \href{http://api.phish.net/keys/}{Phish.net API keys}.
First, you'll need to create a username and password. After confirming your
this information, you can follow the link above to create an API key. Once
approved, store the private key somewhere that it is easily retrieved, and
use it in the \code{apikey} argument to each \code{pn_*} function. Alternatively,
you can set a \code{option(phishnet_key = '<your-key'>)} in your .Rprofile.
}
\examples{
\dontrun{
BigCypressNYE <- pn_get_setlist(apikey = "<apikey>", showdate = "1999-12-31")
}

\dontrun{
BigCypressNYE <- get_show_rating(apikey = "<apikey>", showdate = "1999-12-31")
}
\dontrun{
BigCypressNYE <- get_show_notes(apikey = "<apikey>", showdate = "1999-12-31")
}
}
