% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishin_get_songs.R
\name{pi_get_all_songs}
\alias{pi_get_all_songs}
\title{Get All Songs}
\usage{
pi_get_all_songs(per_page = 100)
}
\value{
a data.frame with the slug, names,
original artist, and times played of each song!
}
\description{
The \code{phish.in} API contains a vast amount of information.
The functions here are specific to each type of query that can be sent to the
API - all are prefixed with \code{pi_*}
}
