% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishin_get_date.R
\name{pi_get_show_by_date}
\alias{pi_get_show_by_date}
\title{Get show using the date from the phish.in API}
\usage{
pi_get_show_by_date(date = NULL)
}
\arguments{
\item{date}{The date you wish to get show info for. Format as \code{'YYYY-MM-DD'}}
}
\description{
Get show using the date from the phish.in API
}
