% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishin_get_songs.R
\name{pi_get_songs}
\alias{pi_get_songs}
\title{Query the Phish.in API}
\usage{
pi_get_songs(song)
}
\arguments{
\item{song}{Either the name of a song or \code{"all"}. It appears that the
API does not support requests for more than one song at a time, so looping
over names is required. When \code{song = "all"}, a data frame of song names
and the number of times played is returned. When \code{song = "some song name"},
a data frame of dates played, show song positions, set positions, durations,
number of likes, and a list column of any tags associated with the perforamnce
is returned.

Note that for all Phish.in API calls that include year information,
1983 - 1987 are lumped together as a single entity. Specific dates are still
available for shows and songs played in this time period.}
}
\description{
The \code{phish.in} API contains a vast amount of information.
The functions here are specific to each type of query that can be sent to the
API - all are prefixed with \code{pi_*}
}
