% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishin_get_tracks.R
\name{pi_get_tracks}
\alias{pi_get_tracks}
\title{Query the Phish.in API for all performances of a song}
\usage{
pi_get_tracks(song_name = "Possum", per_page = 100)
}
\arguments{
\item{song_name}{Name of the song, e.g. "Possum"}

\item{per_page}{Results per page (max 100 recommended)}
}
\value{
A data frame with performance details
}
\description{
Returns metadata about every time a given song was performed.
}
