% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishnet_get_setlist.R
\name{pn_get_setlist}
\alias{pn_get_setlist}
\title{Get Setlist}
\usage{
pn_get_setlist(apikey = getOption("phishnet_key"), showdate = NULL)
}
\arguments{
\item{apikey}{key}

\item{showdate}{date in YYYY-MM-DD format}
}
\description{
Get the setlist for one specific phish show using phish.net
}
