% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishnet_get_shownotes.R
\name{pn_get_show_notes}
\alias{pn_get_show_notes}
\title{Get Show Notes}
\usage{
pn_get_show_notes(apikey = getOption("phishnet_key"), showdate = NULL)
}
\description{
Get Show Notes
}
\examples{
\dontrun{
BigCypressNYE <- get_show_notes(apikey = "<apikey>", showdate = "1999-12-31")
}
}
