% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishnet_get_show_rating.R
\name{pn_get_show_rating}
\alias{pn_get_show_rating}
\title{Get Show Rating}
\usage{
pn_get_show_rating(apikey = getOption("phishnet_key"), showdate = NULL)
}
\description{
Get Show Rating
}
\examples{
\dontrun{
BigCypressNYE <- get_show_rating(apikey = "<apikey>", showdate = "1999-12-31")
}
}
