% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_show_rating.R
\name{get_show_rating}
\alias{get_show_rating}
\title{Search for Setlist}
\usage{
get_show_rating(apikey, showdate = NULL)
}
\arguments{
\item{apikey}{apikey}

\item{showdate}{the show setlist in YYYY-MM-DD format}
}
\value{
the selected show's rating
}
\description{
Search for Setlist
}
\examples{
\dontrun{
BigCypressNYE <- get_show_rating(apikey = "<apikey>", showdate = "1999-12-31")
}
}
