% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setlist.R
\name{get_setlist}
\alias{get_setlist}
\title{Search for Setlist}
\usage{
get_setlist(apikey = NULL, showdate = NULL)
}
\arguments{
\item{apikey}{apikey}

\item{showdate}{the show setlist in YYYY-MM-DD format}
}
\value{
the results from the search
}
\description{
Search for Setlist
}
\examples{
\dontrun{
get_setlist(apikey = "<apikey>", showdate = "1999-12-31")
}
}
