% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishin_get_eras.R, R/phishin_get_songs.R,
%   R/phishin_get_tours.R, R/phishin_get_years.R
\name{pi_get_eras}
\alias{pi_get_eras}
\alias{pi_get_song}
\alias{pi_get_tours}
\alias{pi_get_years}
\title{Query the Phish.in API}
\usage{
pi_get_eras(
  apikey = getOption("phishin_key"),
  era = NULL,
  sort_dir = "descending",
  sort_atr = "name",
  per_page = 20,
  page = 1
)

pi_get_song(
  apikey = getOption("phishin_key"),
  song = NULL,
  sort_dir = "descending",
  sort_atr = "name",
  per_page = 20,
  page = 1
)

pi_get_tours(
  apikey = getOption("phishin_key"),
  tour = NULL,
  sort_dir = "descending",
  sort_atr = "name",
  per_page = 20,
  page = 1
)

pi_get_years(
  apikey = getOption("phishin_key"),
  year,
  sort_dir = "descending",
  sort_atr = "name",
  per_page = 20,
  page = 1
)
}
\arguments{
\item{apikey}{Your key for the Phish.net API. This can also be stored
as an option using \code{options('phishin_key') <- 'your_api_key'}. See details
for more information on how to obtain one.'}

\item{era}{The era you wish to get years for}

\item{sort_dir}{direction to sort result in. The direction is used to sort
on \code{sort_atr}.}

\item{sort_atr}{the attribute to sort on. "date", "name", etc}

\item{per_page}{The number of results per page. The API's default is 20, but
this can be set to larger or smaller numbers.}

\item{page}{The page of results to return.}

\item{song}{Either the name of a song or \code{"all"}. It appears that the
API does not support requests for more than one song at a time, so looping
over names is required. When \code{song = "all"}, a data frame of song names
and the number of times played is returned. When \code{song = "some song name"},
a data frame of dates played, show song positions, set positions, durations,
number of likes, and a list column of any tags associated with the perforamnce
is returned.}

\item{tour}{The tour you wish to get information for. Either \code{NULL} or the
name of a tour. It is usually best to get names for a specific tour from a call
using \code{NULL}, as they can be tricky to guess correctly sometimes.}

\item{year}{The desired year. use \code{year = 'all'} to get a show counts
for every year they toured. Enter a specific year to get a data frame with
dates, venue and location data, and complete setlists information for
each show in a year (set list data is the same format as \code{pi_get_song}).}
}
\description{
The \code{phish.in} API contains a vast amount of information.
The functions here are specific to each type of query that can be sent to the
API - all are prefixed with \code{pi_*}
}
\details{
Note that for all Phish.in API calls that include year information,
1983 - 1987 are lumped together as a single entity. Specific dates are still
available for shows and songs played in this time period.
}
